+++ src/Config/Loader/YamlFileLoader.php
@@ -9,7 +9,9 @@ use Symfony\Component\Yaml\Exception\ParseException;
 use Symfony\Component\Yaml\Yaml;
 
 use InvalidArgumentException;
+use function in_array;
 use function is_array;
+use function is_string;
 use function sprintf;
 
 class YamlFileLoader extends FileLoader