+++ CHANGELOG.md
@@ -6,6 +6,15 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 
 ## [Unreleased]
 
+### Changed
+
+* change case of options names
+from [Kebab case](https://en.wikipedia.org/wiki/Letter_case#Special_case_styles) to [Snake case](https://en.wikipedia.org/wiki/Snake_case)
+
+### Fixed
+
+* usage of Symfony OptionsResolver Component in **ConfigurationHandler**
+
 ## [2.0.0-beta.3] - 2019-05-14
 
 ### Added